LLMProvider refactor
This commit introduces a new `LLMProvider` class which serves as a generic interface for interacting with a large language model, such as OpenAI's GPT. It also refactors the application to use this new interface, removing the dependency on OpenAI code from the `MessageThread` class. This helps to separate concerns and will allow for other providers to be added more easily in future.

As part of this refactor:
* Cost estimation has been completely rewritten for improved accuracy and reliability.
* Token usage is now displayed per request, not per session.
* Cost estimation and token usage reporting are now functional for streamed responses. To enable this, the minimum required version of the openai library has been increased.
* The default `temperature` override of 0.6 has been removed, so no temperature value is sent with API requests unless explicitly set.
* API parameters and the current model are now tracked on a per `LLMProvider` (effectively per session at the moment), not per thread, basis.
* API parameters are no longer saved and/or loaded from json files.
* The message attachment json format has been slightly modified, and is backwards incompatible with that used in the previous commit.