Mark Message sticky and attachments fields public and simplify _message_to_openai
After the LLMProvider refactor, it's less important to keep an exact mapping between public fields on Message and the OpenAI API. For consistency, make the sticky and attachments fields public. This slightly changes the serialization format on Message, so add backwards compatibility for reading the v1 format. When working with json files generated by v2 commits before this one, change `_attachments` to `attachments` before loading to preserve serialized attachments.